会員機能完成

+++ app/controllers/public/likes_controller.rb
@@ -1,17 +1,15 @@
 class Public::LikesController < ApplicationController
+ before_action :authenticate_user!
+ 
  def create
     @post = Post.find(params[:post_id])
-    like = current_user.likes.new(post_id: @post.id)
-    like.save
-   
+    current_user.likes.find_or_create_by(post_id: @post.id)
  end
 
   def destroy
     @post = Post.find(params[:post_id])
     like = current_user.likes.find_by(post_id: @post.id)
-    like.destroy
-    
+    like.destroy if like
   end
- 
 end
